# Optimize gesture buffer