--- a/lib/fusuma/plugin/buffers/gesture_buffer.rb
+++ b/lib/fusuma/plugin/buffers/gesture_buffer.rb
@@ -15,11 +15,15 @@ module Fusuma
         def initialize(*args)
           super(*args)
           @cache = {}
+          @cache_select_by = {}
+          @cache_sum10 = {}
         end
 
         def clear
           super.clear
           @cache = {}
+          @cache_select_by = {}
+          @cache_sum10 = {}
         end
 
         def config_param_types
@@ -53,6 +57,8 @@ module Fusuma
 
             @events.delete(e)
             @cache = {}
+            @cache_select_by = {}
+            @cache_sum10 = {}
           end
         end
 
@@ -78,7 +84,7 @@ module Fusuma
         # @param attr [Symbol]
         # @return [Float]
         def sum_last10_attrs(attr) # sums last 10 values of attr (or all if length < 10)
-          cache_entry = ( @cache[[:sum10, attr]] ||= CacheEntry.new(0, 0) )
+          cache_entry = ( @cache_sum10[attr] ||= CacheEntry.new(0, 0) )
           upd_ev = updating_events
           if upd_ev.length > cache_entry.checked + 1 then
             cache_entry.value = upd_ev.last(10).map do |gesture_event|
@@ -134,7 +140,7 @@ module Fusuma
         end
 
         def select_by_type(type)
-          cache_entry = ( @cache[[:select_by, type]] ||= CacheEntry.new(0, self.class.new([])) )
+          cache_entry = ( @cache_select_by[type] ||= CacheEntry.new(0, self.class.new([])) )
           cache_entry.checked.upto(@events.length - 1).each do |i|
             (cache_entry.value.events << @events[i]) if @events[i].record.gesture == type
           end